implement banner along with small other tweaks
