+++ app/data/me.json
@@ -54,7 +54,7 @@
                 "description" : "I led the design and build of a entirely new, scalable, cloud native, battle management command and control system employing a micro-service based, publish-subscribe architecture. This system of systems is capable of automatically planning and executing theatre level military operations in real time. Helped to successfully demonstrate and sell the system as part of a competitive bidding process to an important customer thereby securing funds to continue this critical research."
             },
             {
-                "description" : "I implemented several microservices from scratch in Java and Go forming patterns for other engineers to follow, designed and led the implementation of new feature sets, led daily scrums (for a couple of years), wrote stories in Jira as well as extensive documentation in Confluence, created a release paradigm using Helm, laid out regression testing procedures and spent considerable time mentoring fellow engineers and pushing for change across our software development community."
+                "description" : "I implemented several microservices from scratch in Java and Go establishing patterns for other engineers to follow, designed and led the implementation of new feature sets, led daily scrums (for a couple of years), wrote stories in Jira as well as extensive documentation in Confluence, created a release paradigm using Helm, laid out regression testing procedures and spent considerable time mentoring fellow engineers and pushing for change across our software development community."
             }
             ],
             "skills" : "Java, Go, Python, Spring Boot, Mockito, Docker, Podman, Microservice, RabbitMQ, MongoDB, Redis, Maven, npm, Jenkins, Coverity, Helm, ArgoCD, OpenShift, Rancher, Kubernetes, Domain Driven Design, Trunk-Based Development, Git, Agile, Scrum"
